+++ inst/templates/analyses.R
@@ -1,12 +1,18 @@
 library(tidyverse)
+library(easystats)
+library(snakecase)
 library(benelib)
+library(gt)
+library(brms)
+library(tidybayes)
 
-use_personal_theme()
+use_pomo_theme()
 
 
 
 
 # Data Import -------------------------------------------------------------
+